feat: take out the logic from the input file of the app
