+++ src/data/persons.json
@@ -1,1 +1,1 @@
-[{"id":"9d1752ec-5aef-4a6c-842e-527e4ad9553b","name":"Voltron","age":32,"hobbies":[]},{"id":"790b59f7-2641-43af-b247-a90317de684f","name":"Kat","age":32,"hobbies":["reading","plaing computer games"]},{"id":"be21a864-7e54-4be0-b6fd-3393d771919c","name":"KATOON","age":32,"hobbies":["reading","plaing computer games"]}]
+[{"id":"790b59f7-2641-43af-b247-a90317de684f","name":"Kat","age":32,"hobbies":["reading","plaing computer games"]},{"id":"498a174b-d6d3-44b2-aab0-b44b661e0cfb","name":"Voltron","age":32,"hobbies":[]}]
